requirements/base.txt
fastapi[standard]
sqlalchemy
sqlmodel
pydantic_settings
logger
fastapi_utils
alembic 
passlib 
pyjwt
psycopg2
bcrypt
user-agents
typing-inspect
logtail-python
casbin
casbin_async_sqlalchemy_adapter
redis>=4.2.0
asyncpg
openfga_sdk
boto3
pillow